created ui before playing the game and moved some files to package
