--- src/game/AI.java
@@ -1,6 +1,0 @@
-package game;
-import entity.Entity;
-
-public interface AI {
-    String getNextMove(Entity ghost, Entity pacman, Game game);
-}
