+++ src/game/ai/AI.java
@@ -1,0 +1,7 @@
+package game.ai;
+import entity.Entity;
+import game.Game;
+
+public interface AI {
+    String getNextMove(Entity ghost, Entity pacman, Game game);
+}
