+++ src/ui/GameState.java
@@ -1,0 +1,7 @@
+package ui;
+
+public enum GameState {
+    PLAYING, MENU, EXIT;
+
+    public static GameState state = MENU;
+}
